refactor(ui): one map-style builder, shared
`buildPreviewStyle` is the browser twin of
`services/portal_generator.generate_style`, and it lived inside PortalEditor.
Anything else that needs to draw layers the way a portal draws them — the layer
page coming next — would have had to copy it, and a copy of this function is
the one that drifts: nearly every line encodes a parity decision that took a
bug to learn. Which source a GeoParquet layer draws from. Why an omitted
`fill-outline-color` still paints an edge. Why a hillshade must not be
rescaled. Why points become pillars in 3D.

So it moves to `ui/src/lib/mapStyle.js` as a pure function, and the editor calls
it with its own state. `lonLatBbox` and `rasterTilesUrl` come along, since they
are part of the same job.

MOVED, not rewritten. The body is byte-identical apart from eighteen lines: the
signature, `basemap` as a parameter instead of a call, the draw-order line now
that the caller supplies the order, and three `dataStore.*` lookups becoming
arguments. Verified by diffing the extracted body against the original — 195 of
213 lines unchanged.

Checked for the failure this refactor invites: no identifier in the module is
undefined, no editor state leaks into it, and nothing the editor still uses was
left behind. `node --check` cannot see any of that, so it was checked
explicitly.

Behaviour is unchanged by construction, but this is the editor's preview path:
worth a look at a portal preview after deploying, not only at the new page.
